Moved Mainwindow gui setup to ui file
